requirements.txt
attrs==24.3.0
exceptiongroup==1.2.2
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.0.2
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.36.1
rpds-py==0.22.3
six==1.17.0
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
